+++ proto/control-api/build.rs
@@ -3,11 +3,13 @@ use std::{error::Error, fs::File, io::ErrorKind};
 #[cfg(feature = "grpc")]
 fn main() -> Result<(), Box<dyn Error>> {
     const GRPC_DIR: &str = "src/grpc/";
-    const GRPC_SPEC_FILE: &str = "src/grpc/control_api.proto";
-    const OUT_FILES: [&str; 2] =
-        ["src/grpc/control_api.rs", "src/grpc/control_api_grpc.rs"];
+    const GRPC_SPEC_FILE: &str = "src/grpc/api.proto";
+    const OUT_FILES: [&str; 2] = ["src/grpc/api.rs", "src/grpc/api_grpc.rs"];
 
-    println!("cargo:rerun-if-changed={}", GRPC_DIR);
+    println!("cargo:rerun-if-changed={}", GRPC_SPEC_FILE);
+    for filename in &OUT_FILES {
+        println!("cargo:rerun-if-changed={}", filename);
+    }
 
     for filename in &OUT_FILES {
         if let Err(e) = File::open(filename) {
@@ -18,10 +20,10 @@ fn main() -> Result<(), Box<dyn Error>> {
                     &GRPC_DIR,
                     None,
                 )
-                .expect("Failed to compile gRPC definitions!");
+                .expect("Failed to compile gRPC definitions");
                 break;
             } else {
-                panic!("{:?}", e);
+                panic!("{}", e);
             }
         }
     }