+++ proto/control-api/src/lib.rs
@@ -3,5 +3,6 @@
 //! [Medea]: https://github.com/instrumentisto/medea
 //! [Control API]: https://tinyurl.com/yxsqplq7
 
+#[rustfmt::skip]
 #[cfg(feature = "grpc")]
 pub mod grpc;