+++ src/api/control/room.rs
@@ -6,7 +6,7 @@ use std::{collections::HashMap, convert::TryFrom};
 
 use derive_more::{Display, From};
 #[rustfmt::skip]
-use medea_control_api_proto::grpc::control_api::{
+use medea_control_api_proto::grpc::api::{
     CreateRequest_oneof_el as ElementProto
 };
 use serde::Deserialize;