+++ src/signalling/room_service.rs
@@ -8,7 +8,7 @@ use actix::{
 use derive_more::Display;
 use failure::Fail;
 use futures::future::{self, Future};
-use medea_control_api_proto::grpc::control_api::Element as ElementProto;
+use medea_control_api_proto::grpc::api::Element as ElementProto;
 
 use crate::{
     api::control::{