+++ proto/client-api/CHANGELOG.md
@@ -13,7 +13,7 @@ All user visible changes to this project will be documented in this file. This p
 
 - `TrackId` and `PeerId` types ([#28]);
 - `Incrementable` trait ([#28]);
-- `RpcConnectionClosed` and `CloseDescription` ([#33](/../../pull/28)).
+- `RpcConnectionClosed` and `CloseDescription` types ([#33](/../../pull/28)).
 
 [#28]: /../../pull/28